requirements.txt
fastapi
uvicorn
openai
google-genai
groq
requests
pydantic